Fix all features build
